# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProsOnWork",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "local server"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Retrieve a list of products",
                "operationId": "listProducts",
                "parameters": [
                    {
                        "name": "longitude",
                        "in": "query",
                        "description": "Longitude of the location for filtering products",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "latitude",
                        "in": "query",
                        "description": "Latitude of the location for filtering products",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search query for filtering products",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "rating",
                        "in": "query",
                        "description": "Rating filter, can accept multiple values",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    {
                        "name": "categories",
                        "in": "query",
                        "description": "Filter by categories, can accept multiple values",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    {
                        "name": "minPrice",
                        "in": "query",
                        "description": "Minimum price for filtering products",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "maxPrice",
                        "in": "query",
                        "description": "Maximum price for filtering products",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "operationId": "storeProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Retrieve a specific product",
                "operationId": "getProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product deleted"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreProductRequest": {},
...
            "UpdateProductRequest": {},
...
            "CategoryResource": {
                "description": "Represents a category",
                "required": [
                    "id",
                    "name",
                    "icon",
                    "image"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Electronics"
                    },
                    "icon": {
                        "$ref": "#/components/schemas/MediaResource"
                    },
                    "image": {
                        "$ref": "#/components/schemas/MediaResource"
                    }
                },
                "type": "object"
            },
...
            "MediaResource": {
                "description": "Represents a media file",
                "required": [
                    "id",
                    "filename",
                    "filePath"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "filename": {
                        "type": "string",
                        "example": "image.png"
                    },
                    "filePath": {
                        "type": "string",
                        "example": "/storage/media/image.png"
                    },
                    "fileType": {
                        "type": "string",
                        "example": "image/png"
                    },
                    "size": {
                        "type": "integer",
                        "example": 204800
                    }
                },
                "type": "object"
            },
            "ProductResource": {
                "required": [
                    "id",
                    "serviceProvider",
                    "category",
                    "name",
                    "description",
                    "address",
                    "border",
                    "isActive",
                    "price",
                    "priceType",
                    "type",
                    "includes",
                    "faq",
                    "images",
                    "videos"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "serviceProvider": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "category": {
                        "$ref": "#/components/schemas/CategoryResource"
                    },
                    "name": {
                        "type": "string",
                        "example": "Web Development Service"
                    },
                    "description": {
                        "type": "string",
                        "example": "Custom website development for businesses."
                    },
                    "address": {
                        "$ref": "#/components/schemas/Address"
                    },
                    "border": {
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "isActive": {
                        "type": "boolean",
                        "example": true
                    },
                    "price": {
                        "type": "number",
                        "format": "double",
                        "example": 49.99
                    },
                    "priceType": {
                        "$ref": "#/components/schemas/PriceTypeEnum"
                    },
                    "type": {
                        "$ref": "#/components/schemas/ProductTypeEnum"
                    },
                    "includes": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Hosting",
                            "Domain"
                        ]
                    },
                    "faq": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FaqResource"
                        }
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MediaResource"
                        }
                    },
                    "videos": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MediaResource"
                        }
                    }
                },
                "type": "object"
            },
            "FaqResource": {
                "properties": {
                    "question": {
                        "type": "string",
                        "example": "Do you provide support?"
                    },
                    "answer": {
                        "type": "string",
                        "example": "Yes, we provide 24/7 support."
                    }
                },
                "type": "object"
            },
            "PriceTypeEnum": {
                "description": "different pricing types",
                "type": "string",
                "enum": [
                    "HOUR",
                    "PACKAGE"
                ],
                "example": "HOUR"
            },
            "ProductTypeEnum": {
                "description": "different product types",
                "type": "string",
                "enum": [
                    "SERVICE",
                    "EQUIPMENT"
                ],
                "example": "SERVICE"
            },
            "UserResource": {
                "required": [
                    "id",
                    "name",
                    "email",
                    "role",
                    "address"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "role": {
                        "$ref": "#/components/schemas/UserRoleEnum"
                    },
                    "address": {
                        "$ref": "#/components/schemas/Address"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "avatar": {
                        "$ref": "#/components/schemas/MediaResource"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "UserRoleEnum": {
                "description": "Possible roles for a user",
                "type": "string",
                "enum": [
                    "CLIENT",
                    "SERVICE_PROVIDER",
                    "ADMIN"
                ]
            },
            "Address": {
                "required": [
                    "longitude",
                    "latitude"
                ],
                "properties": {
                    "longitude": {
                        "type": "number",
                        "format": "double"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}